Updates for MAVLab course 2019 (#13)
* update makefile to build for parrot as well as for pc with BUILD macro

* update to 3.4

* update debug build

* WIP SEGFAULT compile 3.4 for course bebop and pc

* WIP Copy Makefile -D's from 3.2.0 version (does not fix segfault)

* Update opencv version 3.4.0 -> 3.4.5

* WIP Get compilation working for nps

* WIP clean up makefile

* Reduce OpenCV NPS dependencies

* Restructure makefile

Rewrite the Makefile targets to support building multiple targets
at once.

Note that this commit renames the 'install' folder for arm binaries
to 'install_arm', this should also be changed in Paparazzi.

* Minor makefile changes

* Fix link.py, update readme

* Add opencv_contrib

Added opencv_contrib as git submodule. Changed Makefile to
include these in OpenCV compilation.

Tested in NPS with the cv_opencvdemo module (not with SURF or other
contrib features, though...), for ARM only verified that OpenCV
compiles without errors.

WIP del opencv_contrib *subfolder*

FIX opencv_contrib is now submodule

* FIX set opencv_contrib back to 3.4 version

* Remove `make cs` from .travis.yml

`make cs` is no longer a valid target. The default `make all` already compiles both for arm and pc.

* Change travis.yml compiler to `g++`

Current build fails because compiler `arm-linux-none-eabi-g++` is not found when building target `pc`. The target pc uses the default compiler, which should be gcc rather than the cross-compiler. The makefile explicitly sets the compiler when building for arm, so this step should not be affected.

* .travis.yml gcc instead of g++